requirements.txt
discord.py
asyncio.py